pyproject.toml
[project]
name = "arox"
version = "0.0.1"
authors = [
    {name="fangzhen", email="fangzhen@live.com"},
]
description = "Flexible LLM-based Agents"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "tomli",
    "prompt-toolkit>=3.0",
    "pyyaml>=6.0",
    "kissllm",
    "asyncio",
    "GitPython>=3.1.30",
    "networkx[default]>=3.5",
    "grep-ast>=0.9.0",
    "diskcache>=5.6.3",
    "pygments>=2.19.1",
    "tqdm>=4.67.1",
    "tree-sitter-language-pack>=0.7.3",
    "pytest-asyncio>=1.0.0",
]
[project.urls]
Homepage = "https://github.com/Arocial/arox"

[project.scripts]
arox-coder = "arox.compose.coder.main:main"

[tool.setuptools.packages.find]
where = ["."]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
